+++ src/die.js
@@ -1,6 +1,5 @@
 export class TensionDie extends Die {
     constructor(termData) {
-        console.log(JSON.stringify(termData));
         termData.faces=6;
         super(termData);
     }